+++ Engine/Source/GUI/Components/ScrollView.h
@@ -6,7 +6,7 @@ namespace TextureGenEngine
     class ScrollView : public Panel
     {
     public:
-        TAPI ScrollView(int x, int y, int width, int height, ScalingType xScaling, ScalingType yScaling);
+        TAPI ScrollView(float x, float y, float width, float height, ScalingType xScaling, ScalingType yScaling);
         void Draw() override;
         ~ScrollView();
     };
